Instantiation and Addition Fixes
If tool.instantiate called directly, the tool will attempt to find the best Python base class to use.

For methods on data objects, there is now the option to pass a URI instead of an object (in most cases). The tool will attempt to determine the base class and create a reference wrapper object if the object is not currently in the store